feat: add Tabs component
