+++ src/components/Tabs/Tabs.jsx
@@ -1,0 +1,40 @@
+import './Tabs.scss'
+import { useState } from "react";
+
+export const TabItem = props => <div {...props} />;
+
+export const Tabs = ({ defaultIndex = '0', onTabClick, children }) => {
+  const [bindIndex, setBindIndex] = useState(defaultIndex);
+  const changeTab = newIndex => {
+    if (typeof onTabClick === 'function') onTabClick(newIndex);
+    setBindIndex(newIndex);
+  };
+  const items = children.filter(item => item.type.name === 'TabItem');
+
+  return (
+    <div className="tab-menu__wrapper">
+      <div className="tab-menu">
+        {items.map(({ props: { index, label } }) => (
+          <button
+            key={`tab-btn-${index}`}
+            onClick={() => changeTab(index)}
+            className={bindIndex === index ? 'focus' : ''}
+          >
+            {label}
+          </button>
+        ))}
+      </div>
+      <div className="tab-view">
+        {items.map(({ props }) => (
+          <div
+            {...props}
+            className={`tab-content ${
+              bindIndex === props.index ? 'selected' : ''
+            }`}
+            key={`tab-content-${props.index}`}
+          />
+        ))}
+      </div>
+    </div>
+  );
+};
