+++ src/components/Tabs/index.tsx
@@ -1,0 +1,3 @@
+import { Tabs, TabItem } from "./Tabs";
+
+export { TabItem, Tabs }
