+++ src/routes/Profile/index.tsx
@@ -1,3 +1,1 @@
-import {Profile, useUserData} from "./Profile";
-
-export {Profile, useUserData};
+export { default as Profile} from "./Profile";
