+++ src/routes/index.ts
@@ -11,9 +11,10 @@ import Login from './Login';
 import { PrivacyPolicy } from './PrivacyPolicy';
 import { RecipePage } from './RecipePage';
 import { ProgramPage } from './ProgramPage';
+import { Profile } from './Profile';
 
 export {
     Error, Main, Home, WorkoutPrograms,
-    WorkoutVideos, Recipes, About,
+    WorkoutVideos, Recipes, About, Profile,
     Login, Signup, PrivacyPolicy, RecipePage, ProgramPage
 }
